+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Enhancements
 1. Support generating a JSON Schema title from a JSON payload (#505)
 2. Add support for saving Azure key version with DEK (#507)
+3. Pass context when clients make KEK calls to DEK Registry (#508)
 
 
 # confluent-kafka-javascript 1.10.0